pull out defaults and switcher into their own file
