+++ src/docs/spec/default-props.js
@@ -1,0 +1,12 @@
+const addDefaultValues = propData => {
+  /* filter out internal props, they start with _underscore */
+  const propNames = Object.keys(propData).filter(key => key[0] !== '_')
+
+  propNames.forEach(name => {
+    const { defaultValue, type } = propData[name]
+    propData[name].value = defaultValue ? defaultValue.value : 'null'
+  })
+  return propData
+}
+
+export default addDefaultValues
