+++ src/docs/spec/prop-switcher.js
@@ -1,0 +1,18 @@
+import React from 'react'
+import { TextInput, Switch, Select } from '../../components'
+
+const PropSwitcher = ({ propName, data, onPropsChange }) => {
+  let method = value => onPropsChange(propName, value.toString())
+
+  if (data.type.name === 'bool') {
+    return <Switch accessibleLabels={[]} on={data.value === 'true'} onToggle={method} />
+  } else if (['string', 'number'].includes(data.type.name)) {
+    return <TextInput defaultValue={data.value} onChange={e => method(e.target.value)} />
+  } else if (data.type.name === 'enum') {
+    const options = data.type.value.map(({ value }) => ({ text: value, value }))
+    return <Select onChange={e => method(e.target.value)} options={options} />
+  }
+  return <div />
+}
+
+export default PropSwitcher
