+++ src/docs/spec/props.js
@@ -2,8 +2,10 @@ import React from 'react'
 import styled from 'styled-components'
 
 import { spacing, colors, fonts, misc } from '../../tokens'
-import { TextInput, Switch, Code, Select } from '../../components'
+import { Code } from '../../components'
 import parseType from './prop-type'
+import addDefaultValues from './default-props'
+import PropSwitcher from './prop-switcher'
 
 const Table = styled.table`
   width: 100%;
@@ -44,35 +46,12 @@ const Description = styled.span`
   color: ${colors.base.grayDark};
 `
 
-const PropSwitcher = ({ propName, data, onPropsChange }) => {
-  let method = value => onPropsChange(propName, value.toString())
-
-  if (data.type.name === 'bool') {
-    return <Switch accessibleLabels={[]} on={data.value === 'true'} onToggle={method} />
-  } else if (['string', 'number'].includes(data.type.name)) {
-    return <TextInput defaultValue={data.value} onChange={e => method(e.target.value)} />
-  } else if (data.type.name === 'enum') {
-    const options = data.type.value.map(({ value }) => ({ text: value, value }))
-    return <Select onChange={e => method(e.target.value)} options={options} />
-  }
-  return <div />
-}
-
 class Props extends React.Component {
   constructor(props) {
     super(props)
-    const propData = this.getDefaultValues(props.propData)
+    const propData = addDefaultValues(props.propData)
     this.state = { propData: propData }
     this.props.onPropsChange(propData)
-  }
-
-  getDefaultValues(propData) {
-    const propNames = Object.keys(propData).filter(key => key[0] !== '_')
-    propNames.forEach(name => {
-      const { defaultValue } = propData[name]
-      propData[name].value = defaultValue ? defaultValue.value : 'null'
-    })
-    return propData
   }
 
   onPropsChange(propName, value) {
